Add basic LSTM implementation
This is based on https://colah.github.io/posts/2015-08-Understanding-LSTMs/.

Now I can evaluate one pass, but it needs something to drive evaluation
and propagate the values from one step to the next. And it still lacks
the hardest part: actual learning, backpropagating error through the
network and correcting the weights.