Learn on word list, print predictions
These predictions are actually kind of funny. But learning does not seem
to work so well. The error converges quickly, but to a value which is
far too high. Is my network too small, or am I doing something wrong? It
seems to me, that as it contains a 26x26 matrix anyway, it should be
able to just learn the matrix of transition probabilities, and that
should already work reasonably well.