Make it a two-cell network
This appears to work better than the one-cell network, but I haven't run
it long enough to converge to something reasonable. With 10 values
between the two layers, at least it converges to an error similar to the
previous one (around 0.85 per letter) more quickly than the single-cell
network does. But it does not seem to go much further than that. I am
trying with 15 values now, and a decreased learning rate to avoid
overshooting a local minimum. The results look promising, but it is very
slow, so I cannot draw any conclusions yet. I should Rewrite It In Rust
...